Factor out the “config” concern from the salesforce module.
