+++ export.go
@@ -7,6 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/heroku/force/project"
 	"github.com/heroku/force/salesforce"
 	"github.com/heroku/force/util"
 )
@@ -148,7 +149,7 @@ func runExport(cmd *Command, args []string) {
 		fmt.Printf("Encountered and error with retrieve...\n")
 		util.ErrorAndExit(err.Error())
 	}
-	root, err = salesforce.GetSourceDir()
+	root, err = project.GetSourceDir()
 	if err != nil {
 		fmt.Printf("Error obtaining root directory\n")
 		util.ErrorAndExit(err.Error())
