+++ logins.go
@@ -8,6 +8,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/heroku/force/salesforce"
+	"github.com/heroku/force/util"
 )
 
 var cmdLogins = &Command{
@@ -25,7 +26,7 @@ Examples:
 
 func runLogins(cmd *Command, args []string) {
 	active, _ := ActiveLogin()
-	accounts, _ := salesforce.Config.List("accounts")
+	accounts, _ := util.Config.List("accounts")
 	if len(accounts) == 0 {
 		fmt.Println("no logins")
 	} else {
@@ -35,7 +36,7 @@ func runLogins(cmd *Command, args []string) {
 		for _, account := range accounts {
 			if !strings.HasPrefix(account, ".") {
 				var creds salesforce.ForceCredentials
-				data, err := salesforce.Config.Load("accounts", account)
+				data, err := util.Config.Load("accounts", account)
 				json.Unmarshal([]byte(data), &creds)
 
 				if err != nil {
@@ -58,9 +59,9 @@ func runLogins(cmd *Command, args []string) {
 }
 
 func ActiveLogin() (account string, err error) {
-	account, err = salesforce.Config.Load("current", "account")
+	account, err = util.Config.Load("current", "account")
 	if err != nil {
-		accounts, _ := salesforce.Config.List("accounts")
+		accounts, _ := util.Config.List("accounts")
 		if len(accounts) > 0 {
 			SetActiveLoginDefault()
 		}
@@ -73,7 +74,7 @@ func ActiveCredentials() (creds salesforce.ForceCredentials, err error) {
 	if err != nil {
 		return
 	}
-	data, err := salesforce.Config.Load("accounts", account)
+	data, err := util.Config.Load("accounts", account)
 	json.Unmarshal([]byte(data), &creds)
 
 	return
@@ -89,7 +90,7 @@ func ActiveForce() (force *salesforce.Force, err error) {
 }
 
 func SetActiveLoginDefault() (account string) {
-	accounts, _ := salesforce.Config.List("accounts")
+	accounts, _ := util.Config.List("accounts")
 	if len(accounts) > 0 {
 		account = accounts[0]
 		SetActiveLogin(account)
@@ -98,6 +99,6 @@ func SetActiveLoginDefault() (account string) {
 }
 
 func SetActiveLogin(account string) (err error) {
-	err = salesforce.Config.Save("current", "account", account)
+	err = util.Config.Save("current", "account", account)
 	return
 }
