+++ util/config.go
@@ -1,0 +1,7 @@
+package util
+
+import "github.com/devangel/config"
+
+//"fmt"
+
+var Config = config.NewConfig("force")
